+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.22.0]
+- Fix typo in `Formatter`
+- Include some formatters
+- Add `StaticDataFetcher`
+- Change on actions column (Only `ActionsFormatter`)
+
 ## [0.21.10]
 - Remove not needed `getLimitEnd` and rename `getLimitStart` to `getOffset`